schonsteil/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
cycler==0.11.0
defusedxml==0.7.1
Django==4.2
django-cors-headers==3.14.0
django-extensions==3.2.1
django-filter==23.1
django-multiselectfield==0.1.12
django-pictures==1.0.1
django-templated-mail==1.1.1
django-tinymce==3.6.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-spectacular==0.26.2
fonttools==4.39.3
frozenlist==1.3.3
geographiclib==2.0
geoip2==4.7.0
geopy==2.3.0
gpx-cmd-tools==0.0.1
gpxpy==1.5.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
maxminddb==2.3.0
multidict==6.0.4
numpy==1.24.2
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
psycopg2-binary==2.9.6
pycparser==2.21
pydot==1.4.2
PyJWT==2.6.0
pyparsing==3.0.9
pyproj==3.5.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
shapely==2.0.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.1
sorl-thumbnail==12.9.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.15
yarl==1.9.2
